Support passing Env parameters for the running of a guide
We already had this for presteps, required for the actual running of a
guide itself too
